requirements.txt
python-chess
tabulate